+++ front/pages/api/v1/w/[wId]/data_sources/[dsId]/documents/[documentId]/parents.ts
@@ -1,152 +1,7 @@
-import type { WithAPIErrorResponse } from "@dust-tt/types";
-import { CoreAPI } from "@dust-tt/types";
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import config from "@app/lib/api/config";
-import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
-import type { Authenticator } from "@app/lib/auth";
-import { DataSourceResource } from "@app/lib/resources/data_source_resource";
-import logger from "@app/logger/logger";
-import { apiError } from "@app/logger/withlogging";
-
-export type PostParentsResponseBody = {
-  updated: true;
-};
+import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/documents/[documentId]/parents";
 
 /**
- * @swagger
- * /api/v1/w/{wId}/data_sources/{dsId}/documents/{documentId}/parents:
- *   post:
- *     summary: Update the parents of a document
- *     description: Update the parents of a document in the data source identified by {dsId} in the workspace identified by {wId}.
- *     tags:
- *       - Datasources
- *     parameters:
- *       - in: path
- *         name: wId
- *         required: true
- *         description: Unique string identifier for the workspace
- *         schema:
- *           type: string
- *       - in: path
- *         name: dsId
- *         required: true
- *         description: ID of the data source
- *         schema:
- *           type: string
- *       - in: path
- *         name: documentId
- *         required: true
- *         description: ID of the document
- *         schema:
- *           type: string
- *     security:
- *       - BearerAuth: []
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               parents:
- *                 type: array
- *                 items:
- *                   type: string
- *                 description: Array of parent document IDs
- *     responses:
- *       200:
- *         description: The parents were updated
- *       400:
- *         description: Bad Request. Missing or invalid parameters.
- *       401:
- *         description: Unauthorized. Invalid or missing authentication token.
- *       500:
- *         description: Internal Server Error.
- *       404:
- *         description: Data source or workspace not found.
- *       405:
- *         description: Method not supported.
+ * @ignoreswagger
+ * Legacy endpoint.
  */
-
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<PostParentsResponseBody>>,
-  auth: Authenticator
-): Promise<void> {
-  const { dsId } = req.query;
-  if (typeof dsId !== "string") {
-    return apiError(req, res, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Invalid path parameters.",
-      },
-    });
-  }
-
-  const dataSource = await DataSourceResource.fetchByNameOrId(
-    auth,
-    dsId,
-    // TODO(DATASOURCE_SID): Clean-up
-    { origin: "v1_data_sources_documents_document_parents" }
-  );
-  if (!dataSource) {
-    return apiError(req, res, {
-      status_code: 404,
-      api_error: {
-        type: "data_source_not_found",
-        message: "The data source you requested was not found.",
-      },
-    });
-  }
-
-  switch (req.method) {
-    case "POST":
-      if (
-        !req.body ||
-        !Array.isArray(req.body.parents) ||
-        !req.body.parents.every((p: any) => typeof p === "string")
-      ) {
-        return apiError(req, res, {
-          status_code: 400,
-          api_error: {
-            type: "invalid_request_error",
-            message: "Invalid request body, `parents` (string[]) is required.",
-          },
-        });
-      }
-      const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
-      const updateRes = await coreAPI.updateDataSourceDocumentParents({
-        projectId: dataSource.dustAPIProjectId,
-        dataSourceId: dataSource.dustAPIDataSourceId,
-        documentId: req.query.documentId as string,
-        parents: req.body.parents,
-      });
-
-      if (updateRes.isErr()) {
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: "There was an error updating the `parents` field.",
-            data_source_error: updateRes.error,
-          },
-        });
-      }
-
-      res.status(200).json({ updated: true });
-      return;
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, POST is expected.",
-        },
-      });
-  }
-}
-
-export default withPublicAPIAuthentication(handler);
+export default handler;
